+++ tests/OuterListTest.php
@@ -6,6 +6,7 @@ namespace Bakame\Http\StructuredFields;
 
 use LogicException;
 use PHPUnit\Framework\Attributes\Test;
+use TypeError;
 
 final class OuterListTest extends StructuredFieldTestCase
 {
@@ -103,7 +104,7 @@ final class OuterListTest extends StructuredFieldTestCase
     #[Test]
     public function it_fails_to_insert_something_other_than_a_inner_list_or_an_item(): void
     {
-        $this->expectException(InvalidArgument::class);
+        $this->expectException(TypeError::class);
 
         OuterList::new()->push(Dictionary::fromAssociative(['foo' => 'bar']));
     }